feat: Add code to find all subarrays with a given sum
This commit adds a new function, `findAllSubarraysWithGivenSum`, to the code. This function takes a vector of integers and a target sum as input and returns the number of subarrays in the vector that have the given sum. The function uses nested loops to iterate through all possible subarrays and calculates their sum. If the sum matches the target sum, the count is incremented. The main function is also updated to demonstrate the usage of the new function.